refactor: move oauth module to gateway/api + move sandbox module (#96)
