+++ desktop/src-tauri/src/gateway/api/mod.rs
@@ -5,6 +5,7 @@ pub mod chat;
 pub mod external_mcp_client;
 pub mod mcp_request_log;
 pub mod mcp_server;
+pub mod oauth;
 
 pub fn create_router(db: DatabaseConnection) -> Router {
     Router::new()
