+++ desktop/src-tauri/src/lib.rs
@@ -7,7 +7,7 @@ pub mod gateway;
 pub mod models;
 pub mod ollama;
 pub mod openapi;
-pub mod utils;
+pub mod sandbox;
 
 #[cfg(test)]
 pub mod test_fixtures;
@@ -33,11 +33,8 @@ pub fn run() {
                     debug!("SINGLE INSTANCE: Found deep link in argv: {arg}");
                     let app_handle = app.clone();
                     tauri::async_runtime::spawn(async move {
-                        models::mcp_server::oauth::handle_oauth_callback(
-                            app_handle,
-                            arg.to_string(),
-                        )
-                        .await;
+                        gateway::api::oauth::handle_oauth_callback(app_handle, arg.to_string())
+                            .await;
                     });
                 }
             }
@@ -62,8 +59,7 @@ pub fn run() {
             // Start all persisted MCP servers
             let app_handle = app.handle().clone();
             tauri::async_runtime::spawn(async move {
-                if let Err(e) = models::mcp_server::sandbox::start_all_mcp_servers(app_handle).await
-                {
+                if let Err(e) = sandbox::start_all_mcp_servers(app_handle).await {
                     error!("Failed to start MCP servers: {e}");
                 }
             });
@@ -110,11 +106,8 @@ pub fn run() {
                     debug!("DEEP LINK PLUGIN: Processing URL: {url}");
                     let app_handle = app_handle.clone();
                     tauri::async_runtime::spawn(async move {
-                        models::mcp_server::oauth::handle_oauth_callback(
-                            app_handle,
-                            url.to_string(),
-                        )
-                        .await;
+                        gateway::api::oauth::handle_oauth_callback(app_handle, url.to_string())
+                            .await;
                     });
                 }
             });
